fixed a blue hue species issue with sim freezing

+++ neuralnetwork/util.js
@@ -8,7 +8,7 @@ const relativeLeft = (heading, vectAngle) => (heading < vectAngle ? heading + (2
 
 const relativeRight = (heading, vectAngle) => heading < vectAngle ? vectAngle - heading : vectAngle + (2 * Math.PI - heading);
 
-const randomBlueHue = () => randomInt(31) + 225;
+const randomBlueHue = () => randomFloat(31) + 225;
 
 const topoSort = (nodes, edges) => {
     let inMap = new Map();
